update name
update run config
