+++ plugin-build/gradle.properties
@@ -1,8 +1,11 @@
-ID=io.github.frankois944.spmForKmp.plugin
+ID=io.github.frankois944.spmForKmp
 VERSION=0.0.1
-GROUP=io.github.frankois944.spmForKmp
+ARTIFACT_ID=Spm4Kmp
+GROUP=io.github.frankois944
 DISPLAY_NAME=Swift Package Manager for Kotlin Multiplaform
 DESCRIPTION=A Gradle plugin for integrating Swift packages into a Kotlin multiplatform module
 WEBSITE=https://github.com/frankois944/spm4Kmp
 VCS_URL=https://github.com/frankois944/spm4Kmp
-IMPLEMENTATION_CLASS=io.github.frankois944.spmForKmp.plugin.SpmForKmpPlugin
+IMPLEMENTATION_CLASS=io.github.frankois944.spmForKmp.SpmForKmpPlugin
+SONATYPE_AUTOMATIC_RELEASE=true
+
