+++ plugin-build/plugin/build.gradle.kts
@@ -79,22 +79,11 @@ tasks.create("setupPluginUploadFromEnvironment") {
         val key = System.getenv("GRADLE_PUBLISH_KEY")
         val secret = System.getenv("GRADLE_PUBLISH_SECRET")
 
-        val mavenKey = System.getenv("MAVEN_PUBLISH_KEY")
-        val mavenSecret = System.getenv("MAVEN_PUBLISH_SECRET")
-        val signingKeyId = System.getenv("SIGNING_KEY_ID")
-        val signingKeyPassword = System.getenv("SIGNING_KEY_PASSWORD")
-        val signingFile = System.getenv("SIGNING_FILE")
-
         if (key == null || secret == null) {
             throw GradleException("gradlePublishKey and/or gradlePublishSecret are not defined environment variables")
         }
         System.setProperty("gradle.publish.key", key)
         System.setProperty("gradle.publish.secret", secret)
-        System.setProperty("mavenCentralUsername", mavenKey ?: "")
-        System.setProperty("mavenCentralPassword", mavenSecret ?: "")
-        System.setProperty("signing.keyId", signingKeyId ?: "")
-        System.setProperty("signing.password", signingKeyPassword ?: "")
-        System.setProperty("signing.secretKeyRingFile", signingFile ?: "")
     }
 }
 
@@ -104,7 +93,7 @@ mavenPublishing {
     // Define coordinates for the published artifact
     coordinates(
         groupId = property("GROUP").toString(),
-        artifactId = "SpmForKmp",
+        artifactId = property("ARTIFACT_ID").toString(),
         version = property("VERSION").toString(),
     )
