+++ plugin-build/plugin/src/functionalTest/kotlin/io/github/frankois944/spmForKmp/BasicPackageTest.kt
@@ -1,10 +1,10 @@
-package io.github.frankois944.spmForKmp.plugin
+package io.github.frankois944.spmForKmp
 
 import com.autonomousapps.kit.GradleBuilder.build
 import com.autonomousapps.kit.truth.TestKitTruth.Companion.assertThat
-import io.github.frankois944.spmForKmp.plugin.fixture.KotlinSource
-import io.github.frankois944.spmForKmp.plugin.fixture.SmpKMPTestFixture
-import io.github.frankois944.spmForKmp.plugin.fixture.SwiftSource
+import io.github.frankois944.spmForKmp.fixture.KotlinSource
+import io.github.frankois944.spmForKmp.fixture.SmpKMPTestFixture
+import io.github.frankois944.spmForKmp.fixture.SwiftSource
 import org.junit.jupiter.api.Test
 
 class BasicPackageTest {
